Split out methods for cleaning callers and generating a message. Also support custom logger taking the notifications directly
We have a custom logger, but in order to inspect the callstack and
customize the message, we end up throwing out the notification_str and
reaching into Thread.current[:prosopite_notifications] directly to get
it.

We also have to do our own filtering and cleaning because they haven't
been cleaned yet.

I'm proposing this change to let the customer logger take the
`notifications` keyword that gets the notifications OR a single argument
that is the generated string. In support of that, I clean the callstacks
before calling the custom logger.